lotes e pedidos

+++ src/modules/order_batch/services/create-order-batch.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { RevenuesRepository } from 'src/modules/revenue/repository/contract/RevenuesRepository';
 import { CategoryOrderItemRepository } from 'src/modules/category_order_items/repository/contract/CategoryOrderItemRepository';
 import { RevenuePerClientRepository } from 'src/modules/revenue-per-client/repository/contract/RevenuePerClientRepository';
@@ -28,25 +28,33 @@ export class CreateOrderBatchService {
 
         if (!orderAlReadyExist) {
           this.deleteFile(createOrderBatch.file_invoice_absolute);
-          this.deleteFile(createOrderBatch.file_payment_voucher_absolute);
 
           throw new NotFoundException('Pedido ' + item.fk_order + ' não encontrado.')
         }
 
         var orderSem = orderSemOrderBatch.find((t) => t.id === item.fk_order)
 
-        if (!orderSem) {
+        if (orderSem.fk_orderstatus == "1c69c120002-575f34-1c69-be56-0242ac1201c69" || orderSem.fk_orderstatus == "016b9c84-4e7f-81ee-be56-0242ac1200022fe2af") {
+          this.deleteFile(createOrderBatch.file_invoice_absolute);
+
+          throw new BadRequestException("Para adicionar nota fiscal o pedido deve está com staus 'Revisão Admin' ou 'Entregue' Erro: " + orderSem.numberOrder)
+      }
+
+        if (!orderSem && orderSem.file_invoice != null && orderSem.fk_orderstatus != "016b9c84-4e7f-81ee-be56-0242ac1200022fe2af") {
 
           this.deleteFile(createOrderBatch.file_invoice_absolute);
-          this.deleteFile(createOrderBatch.file_payment_voucher_absolute);
 
-          throw new NotFoundException('Pedido possivelmente está vinculado a um lote - fk_order: ' + item.fk_order);
+          throw new BadRequestException('Pedido possivelmente está vinculado a um lote - fk_order: ' + item.fk_order);
 
         }
       })
     )
 
-    await this.orderBatchRepository.create(createOrderBatch);
+    await this.orderBatchRepository.create(createOrderBatch).then(async () => {
+      createOrderBatch.createOrderBatchItem.map(async (item) => {
+        await this.orderRepository.addInvoiceInOrder(item.fk_order, createOrderBatch.file_invoice, createOrderBatch.invoice_number);
+      })
+    });
 
   }
